Added isInBounds and Npc::getPos2d(), Small fixes

+++ src/globalvars.cpp
@@ -126,7 +126,7 @@ bool debug = false;
 bool inFocus = true;
 bool drawStuffsDone = true;
 sf::View view1(CENTER, HALF_SIZE);
-int airPressure = 2;
+float airPressure = 2;
 sf::Vector2f mousePos;
 int mouseStagnation = 0;
 std::vector<int> selected;
